Add all files

+++ app/Http/Controllers/Training_and_Placement_Cell/StudentFormController.php
@@ -45,23 +45,27 @@ class StudentFormController extends Controller
     public function store(Request $request1, Request8 $request)
     {
     	// $request::input('objective');
-        // $validator = validator::make($request1->all(),[
-        //         'objective'=>'required',
-        //         'qualification1'=>'required',
-        //         'institute1'=>'required',
-        //         'performance1'=>'required|numeric',
-        //         'year1'=>'numeric|max:4'
-        //     ]);
+        $validator = validator::make($request1->all(), [
+            'objective'=>'required|max:255'
+        ]);
+        $validator1 = validator::make($request1->all(),[
+            for($a = 1; $a <= $counter; $a++)
+                'qualification'.$a=>'required',
+                'institute'.$a=>'required',
+                'performance'.$a=>'required|numeric',
+                'year'.$a=>'required|numeric|max:4'
+            
+        ]);
 
-        // if($validator->fails()) {
-        //     return redirect()->back()->withInput()->withErrors($validator);
-        // }
+        if($validator->fails() || $validator1->fails()) {
+            return redirect()->back()->withInput()->withErrors($validator);
+        }
         // foreach($this->request->get('qualification') as $key => $val)
         //  {
         //    $rules['qualification'.$key] = 'required';
         //  }
 
-        // else {
+        else {
             $user = StudentEducation::where('student_id', '=', Auth::user()->username)->get();
             if ($user) {
                DB::table('St_Education')->where('student_id', '=', Auth::user()->username)->delete();
@@ -300,7 +304,7 @@ class StudentFormController extends Controller
                 }
                     
             }
-        // }
+        }
         
         return redirect('/training_and_placement_cell')->with('message', 'Form Submitted Successfully!');
     }
